Split clients models into files to make testing and reading easier
